properly avoid reuploading non-modified files
-u does not avoid reuploading files that haven't been changed since the
last upload. That is the default behavior and can only disabled with
--ignore-times, -I. But for that to work, you need transfer modification
times with -t.

What -u does is skip uploading files that have a newer modified time on
the DESTINATION, which is not something that you commonly need. For
example, if you store translation strings in files, and both you and the
website owner modify them, you can use -u to ensure that you don't
overwrite newer translation files on the server when you upload, without
having to remember to download new changes before every upload. They
would otherwise be overwritten even if you haven't modified your local
files lately.

If you don't transfer the modification times, the versions on the
destination will have the modified time of the upload which is newer
than the last time they were modified locally, and so -u will skip
uploading them because they are newer on the receiver, but it's not
meant to be used for that. You should normally just use -t.

With -u you also risk that if the files on the server have been modified
for some reason after their local counterparts, they won't be updated on
uploads.