Added how to open a terminal and a brief introduction to bash
Before, the page simply assumes you know how to open a terminal.
I get why, but the purpose of the website is stated to be to teach "any
normal person" how to set up a website. Most normal people hardly even
seem to understand what a terminal is (some kind of hacker device?) and
likely don't know how to open one. They could look this up but it's
probably discouraging if you have to, so a short note on how to get a
terminal can't hurt. I also added a (hopefully) brief introduction to
bash. I know readers can just copy and paste commands in the tutorials,
but they will have a hard time doing anything independently if they
don't know the very basics of bash or bourne shells.