Minor Indentation fix - TURN & STUN for calls
Just look at the bottom section. The goal of this is to make it the same as the original. The Hugo version's indentation should work, but it looked off (3 unindented, 1 indented)

Anyway, as minor a PR can be, didn't even use git

P.S. Why did you change code color from green to white? Green was better imo, it was visual indicator that it is indeed code (white is plain text)