updated php, fixed typo, added firewall and docs
This commit updates the PHP version on several articles, which is currently 8.2 on Debian
It also uses symbolic links (such as `php-fpm.sock`) on nginx config files and dependency
packages (such as `php-mbstring`) that will always point to the most recent version of
PHP, so there is no need to change those once PHP is updated.

On places where the PHP version must be specified, such as the
`/etc/php/8.2/` directory and the `php8.2-fpm` systemd service, I have
changed the version accordingly.

I believe I have covered every instance of PHP usage on the website, and
programs should work fine. If I missed something or made a change that
broke the configuration of some program, please let me know.

Additionally, I have fixed a typo on mumble.md, which called the command
`scp` as `spc`, added a command to open the mumble port on your firewall
and added documentation links for further configuration of the server.